src/pc_app/requirements.txt
pyserial>=3.5
Pillow>=9.0